requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
certifi==2025.10.5
deprecation==2.1.0
frozenlist==1.8.0
idna==3.11
magic-filter==1.0.12
multidict==6.7.0
ntgcalls==2.0.6
packaging==25.0
propcache==0.4.1
py-tgcalls==2.2.8
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.11.10
pydantic_core==2.33.2
python-dotenv==1.2.1
rsa==4.9.1
Telethon==1.41.2
typing-inspection==0.4.2
typing_extensions==4.15.0
yarl==1.22.0
yt-dlp==2025.10.22
